fix/remove-erc20-lib (#47)
* we no longer need this lib

* shoved test around

* fix

* rename
